pyproject.toml
[project]
name = "ml-drill"
version = "0.1.0"
description = "ML interview prep tool with RAG and Claude evaluation"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "anthropic>=0.28.0",
    "python-dotenv>=1.0.0",
    "youtube-transcript-api>=1.2.4",
]

[tool.uv]
package = false
